Combine almost identical URL parsers
The Bitbucket and Gitlab URL parsers were almost identical, so it made
sense to just combine them, to avoid extreme code duplication.
