+++ app/models/concerns/bitbucket_project.rb
@@ -1,5 +1,5 @@
 module BitbucketProject
   def bitbucket_name_with_owner
-    BitbucketUrls::Parser.parse(repository_url) || BitbucketUrls::Parser.parse(homepage)
+    BitbucketURLParser.parse(repository_url) || BitbucketURLParser.parse(homepage)
   end
 end
